Moving Tip-related pages to new file structure
